pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mondo"
version = "0.6.0"
description = "Power-user CLI for the monday.com GraphQL API — az/gh/gam style."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "marktguru" }]
keywords = ["monday", "monday.com", "cli", "graphql", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "jmespath>=1.0.1",
    "ruamel.yaml>=0.18.6",
    "keyring>=25.0.0",
    "pydantic>=2.9.0",
    "loguru>=0.7.2",
    "openpyxl>=3.1.5",
    "rapidfuzz>=3,<4",
    "markdown-it-py>=3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-httpx>=0.35.0",
    "syrupy>=4.8.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pyinstaller>=6.11",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mondo = "mondo.cli.main:main"

[project.urls]
Homepage = "https://github.com/zoltanf/mondo"
Issues = "https://github.com/zoltanf/mondo/issues"
Documentation = "https://github.com/zoltanf/mondo#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/mondo"]

[tool.uv]
# uv manages the virtualenv under .venv/ by default.
# Lockfile (uv.lock) is checked in for reproducible builds.

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "UP",       # pyupgrade
    "SIM",      # simplify
    "RUF",      # ruff-specific
]
ignore = [
    "E501",     # line length handled by formatter
]

[tool.ruff.lint.flake8-bugbear]
# Typer/Click expect call-based defaults — these are idiomatic, not bugs.
extend-immutable-calls = ["typer.Option", "typer.Argument"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
files = ["src/mondo"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "integration: integration tests requiring MONDAY_TEST_TOKEN (deselect with '-m \"not integration\"')",
    "eval: skill eval tests requiring ANTHROPIC_API_KEY + MONDO_TEST_BOARD_ID (deselect with '-m \"not eval\"')",
]
